+++ kas-wgpu/src/draw/draw_pipe.rs
@@ -8,18 +8,13 @@
 //! TODO: move traits up to kas?
 
 use std::any::Any;
-use std::borrow::Cow;
 use std::f32::consts::FRAC_PI_2;
+use wgpu_glyph::GlyphBrushBuilder;
 
-use wgpu_glyph::{GlyphBrush, GlyphBrushBuilder, GlyphCruncher, VariedSection};
-
-use super::{Colour, Draw, Vec2};
+use super::{Colour, Draw, DrawPipe, RoundPipe, SquarePipe, Vec2};
+use crate::shared::SharedState;
 use kas::geom::{Coord, Rect, Size};
 use kas::theme;
-
-use super::round_pipe::RoundPipe;
-use super::square_pipe::SquarePipe;
-use crate::shared::SharedState;
 
 /// Style of drawing
 pub enum ShadeStyle {
@@ -56,43 +51,6 @@ pub trait DrawShaded: Draw {
         style: ShadeStyle,
         col: Colour,
     );
-}
-
-/// Abstraction over text rendering
-///
-/// TODO: this API is heavily dependent on `glyph_brush`. Eventually we want our
-/// own API, encapsulating translation functionality and with more default
-/// values (e.g. scale). When we get there, we should be able to move
-/// `SampleTheme` to `kas`.
-pub trait DrawText {
-    /// Queues a text section/layout.
-    fn draw_text<'a, S>(&mut self, section: S)
-    where
-        S: Into<Cow<'a, VariedSection<'a>>>;
-
-    /// Returns a bounding box for the section glyphs calculated using each glyph's
-    /// vertical & horizontal metrics.
-    ///
-    /// If the section is empty or would result in no drawn glyphs will return `None`.
-    ///
-    /// Invisible glyphs, like spaces, are discarded during layout so trailing ones will
-    /// not affect the bounds.
-    ///
-    /// The bounds will always lay within the specified layout bounds, ie that returned
-    /// by the layout's `bounds_rect` function.
-    ///
-    /// Benefits from caching, see [caching behaviour](#caching-behaviour).
-    fn glyph_bounds<'a, S>(&mut self, section: S) -> Option<(Vec2, Vec2)>
-    where
-        S: Into<Cow<'a, VariedSection<'a>>>;
-}
-
-/// Manager of draw pipes and implementor of [`Draw`]
-pub struct DrawPipe {
-    clip_regions: Vec<Rect>,
-    round_pipe: RoundPipe,
-    square_pipe: SquarePipe,
-    glyph_brush: GlyphBrush<'static, ()>,
 }
 
 impl DrawPipe {
@@ -228,25 +186,5 @@ impl DrawShaded for DrawPipe {
             }
             ShadeStyle::Round(norm) => self.round_pipe.shaded_frame(pass, outer, inner, norm, col),
         }
-    }
-}
-
-impl DrawText for DrawPipe {
-    #[inline]
-    fn draw_text<'a, S>(&mut self, section: S)
-    where
-        S: Into<Cow<'a, VariedSection<'a>>>,
-    {
-        self.glyph_brush.queue(section)
-    }
-
-    #[inline]
-    fn glyph_bounds<'a, S>(&mut self, section: S) -> Option<(Vec2, Vec2)>
-    where
-        S: Into<Cow<'a, VariedSection<'a>>>,
-    {
-        self.glyph_brush
-            .glyph_bounds(section)
-            .map(|rect| (Vec2(rect.min.x, rect.min.y), Vec2(rect.max.x, rect.max.y)))
     }
 }
