+++ kas-wgpu/src/draw/draw_text.rs
@@ -1,0 +1,60 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License in the LICENSE-APACHE file or at:
+//     https://www.apache.org/licenses/LICENSE-2.0
+
+//! Text drawing API for `kas_wgpu`
+
+use std::borrow::Cow;
+use wgpu_glyph::{GlyphCruncher, VariedSection};
+
+use super::{DrawPipe, Vec2};
+
+/// Abstraction over text rendering
+///
+/// TODO: this API is heavily dependent on `glyph_brush`. Eventually we want our
+/// own API, encapsulating translation functionality and with more default
+/// values (e.g. scale). When we get there, we should be able to move
+/// `SampleTheme` to `kas`.
+pub trait DrawText {
+    /// Queues a text section/layout.
+    fn draw_text<'a, S>(&mut self, section: S)
+    where
+        S: Into<Cow<'a, VariedSection<'a>>>;
+
+    /// Returns a bounding box for the section glyphs calculated using each glyph's
+    /// vertical & horizontal metrics.
+    ///
+    /// If the section is empty or would result in no drawn glyphs will return `None`.
+    ///
+    /// Invisible glyphs, like spaces, are discarded during layout so trailing ones will
+    /// not affect the bounds.
+    ///
+    /// The bounds will always lay within the specified layout bounds, ie that returned
+    /// by the layout's `bounds_rect` function.
+    ///
+    /// Benefits from caching, see [caching behaviour](#caching-behaviour).
+    fn glyph_bounds<'a, S>(&mut self, section: S) -> Option<(Vec2, Vec2)>
+    where
+        S: Into<Cow<'a, VariedSection<'a>>>;
+}
+
+impl DrawText for DrawPipe {
+    #[inline]
+    fn draw_text<'a, S>(&mut self, section: S)
+    where
+        S: Into<Cow<'a, VariedSection<'a>>>,
+    {
+        self.glyph_brush.queue(section)
+    }
+
+    #[inline]
+    fn glyph_bounds<'a, S>(&mut self, section: S) -> Option<(Vec2, Vec2)>
+    where
+        S: Into<Cow<'a, VariedSection<'a>>>,
+    {
+        self.glyph_brush
+            .glyph_bounds(section)
+            .map(|rect| (Vec2(rect.min.x, rect.min.y), Vec2(rect.max.x, rect.max.y)))
+    }
+}
